==> docs/04_random_forest.Rmd <==
---
output: word_document
---

```{r Global options set, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
source("01_global.R")
```

```{r}
Mixed_2005 <- arrow::read_feather("Tidy_Data/Mixed_Data_2005.feather") |> 
  dplyr::filter(Reference == TRUE, SurveyYear > 2004) |>
  dplyr::mutate(SurveyYear = factor(SurveyYear),
                IslandCode = factor(IslandCode, levels = Island_Code_Levels),
                ReserveStatus = factor(ReserveStatus)) |> 
  dplyr::select(-SiteNumber, -SiteName, 
                -IslandName, -SiteCode) 
RF_Reserve_Model_2005 <- base::readRDS("Models/RF_Reserve_Model_2005.rds")

RF_Importance <- arrow::read_feather("Tidy_Data/RF_Importance.feather") |> 
  dplyr::arrange(desc(MeanDecreaseAccuracy))
partial_df <- arrow::read_feather("Tidy_Data/PDP_data.feather") |> 
  dplyr::mutate(CommonName = fct_inorder(CommonName))
```

## Ran For

```{r RF Variable Importance, message=FALSE, warning=FALSE, fig.height=7.5, fig.width=10}

Top_30_MDA <- RF_Importance %>%
  dplyr::arrange(desc(MeanDecreaseAccuracy)) %>%  
  head(30) %>% 
  droplevels() |> 
  dplyr::mutate(CommonName = forcats::fct_rev(forcats::fct_inorder(CommonName)))

Top_30_MDG <- RF_Importance %>% 
  dplyr::arrange(desc(MeanDecreaseGini)) %>%  
  head(30) %>% 
  droplevels() |> 
  dplyr::mutate(CommonName = forcats::fct_rev(forcats::fct_inorder(CommonName)))

Accuracy <- ggplot2::ggplot(Top_30_MDA, aes(x = MeanDecreaseAccuracy, y = CommonName, color = Targeted)) +
  ggplot2::geom_point() +
  ggplot2::geom_segment(size = 1.25, 
               aes(x = min(MeanDecreaseAccuracy) - 2, xend = MeanDecreaseAccuracy,
                   y = CommonName, yend = CommonName)) +
  ggplot2::geom_hline(yintercept = 20.5, linetype = "dashed") +
  ggplot2::geom_text(size = 5, aes(x = Inf, y = 20.75), show.legend = F, 
                     label = "Top 10 \U1f815", vjust = 0, hjust = 1) +
  ggplot2::geom_hline(yintercept = 10.5, linetype = "dashed") +
  ggplot2::geom_text(size = 5, aes(x = Inf, y = 10.75), show.legend = F, 
                     label = "Top 20 \U1f815", vjust = 0, hjust = 1) +

  ggplot2::labs(x = "Mean Decrease in % Accuracy", y = NULL, tag = "A", 
       color = NULL, linetype = NULL) +
  ggplot2::scale_x_continuous(expand = expansion(mult = c(0,.1)), 
                     limits = c(min(Top_30_MDA$MeanDecreaseAccuracy) - 2, NA)) +
  ggplot2::scale_color_manual(values = Target_Colors, limits = force) +
  ggplot2::theme_classic() +
  ggplot2::theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 12),
        legend.text = element_text(size = 12))

Gini <- ggplot2::ggplot(Top_30_MDG, aes(x = MeanDecreaseGini, y = CommonName, color = Targeted)) +
  ggplot2::geom_point() +
  ggplot2::geom_segment(size = 1.25,
               aes(x = min(MeanDecreaseGini) - 1, xend = MeanDecreaseGini,
                   y = CommonName, yend = CommonName)) +
  ggplot2::geom_hline(yintercept = 20.5, linetype = "dashed") +
  ggplot2::geom_text(size = 5, aes(x = Inf, y = 20.75), show.legend = F, 
                     label = "Top 10 \U1f815", vjust = 0, hjust = 1) +
  ggplot2::geom_hline(yintercept = 10.5, linetype = "dashed") +
  ggplot2::geom_text(size = 5, aes(x = Inf, y = 10.75), show.legend = F, 
                     label = "Top 20 \U1f815", vjust = 0, hjust = 1) +
  ggplot2::labs(x = "Mean Decrease in Gini Index", y = NULL, tag = "B", 
       color = NULL, linetype = NULL) +
  ggplot2::scale_x_continuous(expand = expansion(mult = c(0,.1)), 
                     limits = c(min(Top_30_MDG$MeanDecreaseGini) - 1, NA)) +
  ggplot2::scale_y_discrete() +
  ggplot2::scale_color_manual(values = Target_Colors, limits = force) +
  ggplot2::theme_classic() +
  ggplot2::theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 12),
        legend.text = element_text(size = 12))
ggpubr::ggarrange(Accuracy, Gini, ncol = 2, align = "h", common.legend = TRUE, legend = "bottom")

```

```{r RF PDP Plots, warning=FALSE, message=FALSE, fig.height=8.5, fig.width=7.5}
df_label <- partial_df |> 
  dplyr::distinct(common_name, Rank)

p <- ggplot2::ggplot() +
  ggplot2::geom_smooth(data = partial_df,
                       method = 'loess', formula = 'y ~ x', se = FALSE, size = 1.25, 
                       aes(x = x_var, y = yhat, color = targeted)) + 
  ggplot2::geom_label(data = df_label, label.size = 0,
                      aes(x = -Inf, y = Inf, label = Rank),
                      hjust = 0, vjust = 1, size = 4) +
  # ggplot2::scale_color_manual(values = Target_Colors, limits = force) +
  ggplot2::scale_y_continuous(expand = c(0, 0.01)) +
  ggplot2::scale_x_continuous(expand = c(0, 0)) +
  ggplot2::labs(x = NULL, y = NULL, color = NULL) +
  ggplot2::theme_minimal() +
  PDP_theme()
p1 <- p +
  ggforce::facet_wrap_paginate(facets = vars(common_name), 
                               nrow = 5, ncol = 3, page = 1, scales = "free") 
p2 <- p +
  ggforce::facet_wrap_paginate(facets = vars(common_name), 
                               nrow = 5, ncol = 3, page = 2, scales = "free")
print(p1)
print(p2)
```

```{r RD PDP individual, warning=FALSE, message=FALSE, fig.height=2, fig.width=3}
# Choose any variable to plot PDP

Imp_Num <- 39

pdp <- lapply(RF_Importance$Common_Name[Imp_Num], function(i){
  partial(RF_Reserve_Model_2005, pred.var = i) |>
    data.frame() |>
    dplyr::mutate(Common_Name = i) |>
    dplyr::rename(x_var = i)
})
pdp2 <- pdp
# pdp2[[20]][["x_var"]] <- as.numeric(pdp2[[20]][["x_var"]])
part_df <- bind_rows(pdp2) |>
  dplyr::left_join(Mixed_Data_xRef_Biomass) |> 
  dplyr::mutate(CommonName = paste(CommonName, " (", Data_Type, ")", sep = ""),
                CommonName = gsub(" biomass", "", CommonName),
                CommonName = gsub("Biomass", "B", CommonName),
                CommonName = gsub("Count", "C", CommonName),
                CommonName = gsub("Percent Cover", "P", CommonName),
                CommonName = gsub("Index Value", "I", CommonName),
                CommonName = gsub("Categorical", "Cg", CommonName),
                CommonName = fct_inorder(CommonName)) |>
  dplyr::mutate(Rank = as.integer(CommonName)) 


ggplot2::ggplot() +
  ggplot2::geom_smooth(data = part_df, show.legend = F,
                       method = 'loess', formula = 'y ~ x', se = FALSE, size = 1.25, 
                       aes(x = x_var, y = yhat, color = Targeted)) + 
  ggplot2::geom_label(label.size = 0,
                      aes(x = -Inf, y = Inf, label = Imp_Num),
                      hjust = 0, vjust = 1, size = 4) +
  ggplot2::scale_color_manual(values = Target_Colors, limits = force) +
  ggplot2::scale_y_continuous(expand = c(0, 0.01)) +
  ggplot2::scale_x_continuous(expand = c(0, 0)) +
  ggplot2::labs(x = NULL, y = NULL, color = NULL, title = part_df$CommonName) +
  ggplot2::theme_minimal() +
  PDP_theme() +
  ggplot2::theme(plot.title = element_text(hjust = 0.5, size = 10))
```
