==> docs/11_indicator_sp.Rmd <==
---
output: word_document
---

```{r Global options set, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
source("01_global.R")
```

```{r}
# All_Ratios <- arrow::read_feather("Tidy_Data/Ratios.feather")
```

## Ind Spe Ana

```{r ISA with New Density DF, fig.width=10, fig.height=7.5}

# fishy <- readr::read_csv("../KFM_Stats/Tidy_Data_Dont_Touch/Fish_Counts.csv")
# benny <- readr::read_csv("../KFM_Stats/Tidy_Data_Dont_Touch/Benthic_Counts.csv")
# 
# Denny <- fishy %>%
#     full_join(benny) 

Density_Wide <- Density |> 
  dplyr::filter(!Survey_Type %in% c("VFT", "RPC"),
                
                !ScientificName %in% c(
                  "Alloclinus holderi", "Coryphopterus nicholsi", "Lythrypnus dalli") | 
                  Survey_Type == "1 m² quads",
                !CommonName %in% c(
                  "giant kelp, adult (>1m and haptera above the primary dichotomy)",
                  "giant kelp, subadult (>1m and no haptera above the primary dichotomy)",
                  "Sargassum horneri, juvenile (<50cm and no recepticles)",
                  "Sargassum horneri, adult (>50cm or recepticles present)",
                  "wakame, adult",
                  "wakame, juvenile"
                )) |> 
  dplyr::select(SiteNumber, IslandCode, IslandName, SiteCode, SiteName, SurveyYear, 
                ReserveStatus,
                ScientificName,
                Count) |> 
  tidyr::pivot_wider(names_from = "ScientificName", values_from = "Count", values_fill = 0, values_fn = sum) 

indic_spp_table <- dplyr::tibble(
  ScientificName = character(), s.SR = double(), s.SC = double(), 
  s.AN = double(), s.SB = double(), index = integer(), stat = double(),
  p.value = double(), Year = integer(), ReserveStatus = character())

for (h in base::unique(Density_Wide$ReserveStatus)) {
  for (k in base::unique(Density_Wide$SurveyYear)) {
    Indic_Spp <- Density_Wide %>%
      dplyr::filter(ReserveStatus %in% h,
                    SurveyYear %in% k) %>%
      dplyr::select(-SiteNumber, -IslandCode, -IslandName, 
                    -SiteCode, -SiteName, -SurveyYear, -ReserveStatus) %>%
      indicspecies::multipatt(multipatt_island_list, func = "r.g", control = how(nperm = 1000)) %>% 
      base::getElement('sign') %>%
      tibble::rownames_to_column("ScientificName") %>%
      dplyr::mutate(Year = k, ReserveStatus = h) %>%
      dplyr::filter(p.value < 0.05) 
    indic_spp_table <- base::rbind(indic_spp_table, Indic_Spp)
  }
}

indic_sites <- c("s.SR", "s.SC", "s.AN", "s.SB")

for (h in base::unique(Density_Wide$ReserveStatus)) {
  for (i in indic_sites) {
    Spp_List <- indic_spp_table %>%
      dplyr::filter(!!as.symbol(i) == 1, ReserveStatus %in% h) %>%
      base::getElement('ScientificName') %>%
      base::unique() 
    
    unique_spplist <- base::paste(h, i, "_Spp_List", sep = "")
    base::assign(unique_spplist, Spp_List)
  }
}

heat_map_data <- base::data.frame(
  ScientificName = character(), stat = double(), p.value = double(),
  year = integer(), IslandName = character(), ReserveStatus = character())

for (h in base::unique(Density_Wide$ReserveStatus)) {
  for (z in indic_sites) {
    Spp_Scores <- indic_spp_table %>%
      dplyr::filter(!!as.symbol(z) == 1, ReserveStatus %in% h) %>%
      dplyr::mutate(IslandName = z, ReserveStatus = h)
    heat_map_data <- heat_map_data %>%
      base::rbind(Spp_Scores)
  }
}

heat_map_data <- dplyr::select(
  heat_map_data, ScientificName, stat, p.value, Year, IslandName, ReserveStatus)

Outside_Heat_Species <- 
  base::rep(c(Outsides.SR_Spp_List, Outsides.SC_Spp_List,
              Outsides.AN_Spp_List, Outsides.SB_Spp_List), 
            times = base::length(unique(heat_map_data$Year)))
Inside_Heat_Species <- 
  base::rep(c(Insides.SR_Spp_List, Insides.SC_Spp_List, 
              Insides.AN_Spp_List, Insides.SB_Spp_List), 
            times = base::length(unique(heat_map_data$Year)))
Outside_Heat_Sites <- 
  base::rep(c((base::rep("s.SR", each = length(Outsides.SR_Spp_List))), 
              (base::rep("s.SC", each = base::length(Outsides.SC_Spp_List))), 
              (base::rep("s.AN", each = base::length(Outsides.AN_Spp_List))), 
              (base::rep("s.SB", each = base::length(Outsides.SB_Spp_List)))),
            times = base::length(base::unique(heat_map_data$Year)))
Inside_Heat_Sites <- 
  base::rep(c((base::rep("s.SR", each = length(Insides.SR_Spp_List))), 
              (base::rep("s.SC", each = base::length(Insides.SC_Spp_List))), 
              (base::rep("s.AN", each = base::length(Insides.AN_Spp_List))), 
              (base::rep("s.SB", each = base::length(Insides.SB_Spp_List)))),
            times = base::length(base::unique(heat_map_data$Year)))
Outside_Heat_Year <- 
  base::rep((base::unique(heat_map_data$Year)), each = 
              ((base::length(Outsides.SR_Spp_List))+
                 (base::length(Outsides.SC_Spp_List))+
                 (base::length(Outsides.AN_Spp_List))+
                 (base::length(Outsides.SB_Spp_List))))
Inside_Heat_Year <- 
  base::rep((base::unique(heat_map_data$Year)), each = 
              ((base::length(Insides.SR_Spp_List))+
                 (base::length(Insides.SC_Spp_List))+
                 (base::length(Insides.AN_Spp_List))+
                 (base::length(Insides.SB_Spp_List))))
Outside_full_heat_table <-  base::data.frame(
  IslandName = Outside_Heat_Sites, Year = Outside_Heat_Year, 
  ScientificName = Outside_Heat_Species, ReserveStatus = "Outside")

Inside_full_heat_table <-  base::data.frame(
  IslandName = Inside_Heat_Sites, Year = Inside_Heat_Year, 
  ScientificName = Inside_Heat_Species, ReserveStatus = "Inside")

full_heat_table <- rbind(Outside_full_heat_table, Inside_full_heat_table)

heat_map <- dplyr::left_join(
  full_heat_table, heat_map_data, by=c(
    "IslandName", "Year", "ScientificName", "ReserveStatus")) %>%
  dplyr::mutate(IslandName = dplyr::recode_factor(
    IslandName, s.SR ="Santa Rosa Island", s.SC ="Santa Cruz Island",
    s.AN ="Anacapa Island", s.SB ="Santa Barbara Island")) %>% 
  dplyr::group_by(IslandName, ScientificName, ReserveStatus) %>% 
  dplyr::mutate(frequency = length(stat[!is.na(stat)])) %>% 
  dplyr::ungroup() %>% 
  dplyr::arrange(desc(frequency), desc(ReserveStatus)) 

z <- 1
for(i in levels(heat_map$IslandName)) {
  Island_Heat_Data <- heat_map %>% 
    dplyr::filter(IslandName == i) 
  
  p <- ggplot2::ggplot() +
    geom_tile(data = Island_Heat_Data,
              aes(x = Year, y = ScientificName, fill = stat), color = "black") +
    ggplot2::scale_x_continuous(breaks = 2005:Year_to_Filter_Data_by, expand = c(0, 0)) +
    ggplot2::scale_y_discrete(expand = c(0, 0)) +
    ggplot2::scale_fill_gradient2(low = scales::muted("lightgreen"), 
                                  high = "forestgreen", na.value = scales::muted("firebrick")) + 
    ggplot2::facet_grid(#rows = vars(Classification), 
                        cols = vars(ReserveStatus), scales = "free", space = "free") +
    ggplot2::labs(title = i, x = "Year", y = "Species/Taxa", 
                  fill = paste("Strength of", "\n", "Association")) +
    ISA_theme()
  unique_plotname <- paste("ISA.", z, sep = "")
  assign(unique_plotname, p)
  z <- z + 1
}
ISA.1
ISA.2
ISA.3
ISA.4
```

```{r ISA with top 30, fig.width=10, fig.height=7.5}

Species_to_Use <- RF_Importance |> 
  dplyr::filter(Classification != "Mixed",
                !Data_Type %in% c("Categorical", "Index Value"),
                Targeted != "Calculated Value") |> 
  utils::head(30) |> 
  base::getElement('Common_Name')

ISA_30 <- arrow::read_feather("Tidy_Data/Mixed_Data_2005.feather") |> 
  dplyr::select(SiteNumber, IslandCode, IslandName, SiteName, 
                Species_to_Use,
                SiteCode, SurveyYear, ReserveStatus)

indic_spp_table <- dplyr::tibble(
  ScientificName = character(), s.SR = double(), s.SC = double(), 
  s.AN = double(), s.SB = double(), index = integer(), stat = double(),
  p.value = double(), Year = integer(), ReserveStatus = character())

for (h in base::unique(ISA_30$ReserveStatus)) {
  for (k in base::unique(ISA_30$SurveyYear)) {
    Indic_Spp <- ISA_30 %>%
      dplyr::filter(ReserveStatus %in% h,
                    SurveyYear %in% k) %>%
      dplyr::select(-SiteNumber, -IslandCode, -IslandName, -SiteName, -SiteCode, -SurveyYear, -ReserveStatus) %>%
      indicspecies::multipatt(multipatt_island_list, func = "r.g", control = how(nperm = 1000)) %>% 
      base::getElement('sign') %>%
      tibble::rownames_to_column("ScientificName") %>%
      dplyr::mutate(Year = k, ReserveStatus = h) %>%
      dplyr::filter(p.value < 0.05) 
    indic_spp_table <- base::rbind(indic_spp_table, Indic_Spp)
  }
}

indic_sites <- c("s.SR", "s.SC", "s.AN", "s.SB")

for (h in base::unique(ISA_30$ReserveStatus)) {
  for (i in indic_sites) {
    Spp_List <- indic_spp_table %>%
      dplyr::filter(!!as.symbol(i) == 1, ReserveStatus %in% h) %>%
      base::getElement('ScientificName') %>%
      base::unique() 
    
    unique_spplist <- base::paste(h, i, "_Spp_List", sep = "")
    base::assign(unique_spplist, Spp_List)
  }
}

heat_map_data <- base::data.frame(
  ScientificName = character(), stat = double(), p.value = double(),
  year = integer(), IslandName = character(), ReserveStatus = character())

for (h in base::unique(ISA_30$ReserveStatus)) {
  for (z in indic_sites) {
    Spp_Scores <- indic_spp_table %>%
      dplyr::filter(!!as.symbol(z) == 1, ReserveStatus %in% h) %>%
      dplyr::mutate(IslandName = z, ReserveStatus = h)
    heat_map_data <- heat_map_data %>%
      base::rbind(Spp_Scores)
  }
}

heat_map_data <- dplyr::select(
  heat_map_data, ScientificName, stat, p.value, Year, IslandName, ReserveStatus)

Outside_Heat_Species <- 
  base::rep(c(Outsides.SR_Spp_List, Outsides.SC_Spp_List,
              Outsides.AN_Spp_List, Outsides.SB_Spp_List), 
            times = base::length(unique(heat_map_data$Year)))
Inside_Heat_Species <- 
  base::rep(c(Insides.SR_Spp_List, Insides.SC_Spp_List, 
              Insides.AN_Spp_List, Insides.SB_Spp_List), 
            times = base::length(base::unique(heat_map_data$Year)))
Outside_Heat_Sites <- 
  base::rep(c((base::rep("s.SR", each = base::length(Outsides.SR_Spp_List))), 
              (base::rep("s.SC", each = base::length(Outsides.SC_Spp_List))), 
              (base::rep("s.AN", each = base::length(Outsides.AN_Spp_List))), 
              (base::rep("s.SB", each = base::length(Outsides.SB_Spp_List)))),
            times = base::length(base::unique(heat_map_data$Year)))
Inside_Heat_Sites <- 
  base::rep(c((base::rep("s.SR", each = base::length(Insides.SR_Spp_List))), 
              (base::rep("s.SC", each = base::length(Insides.SC_Spp_List))), 
              (base::rep("s.AN", each = base::length(Insides.AN_Spp_List))), 
              (base::rep("s.SB", each = base::length(Insides.SB_Spp_List)))),
            times = base::length(base::unique(heat_map_data$Year)))
Outside_Heat_Year <- 
  base::rep((base::unique(heat_map_data$Year)), each = 
              ((base::length(Outsides.SR_Spp_List))+
                 (base::length(Outsides.SC_Spp_List))+
                 (base::length(Outsides.AN_Spp_List))+
                 (base::length(Outsides.SB_Spp_List))))
Inside_Heat_Year <- 
  base::rep((base::unique(heat_map_data$Year)), each = 
              ((base::length(Insides.SR_Spp_List))+
                 (base::length(Insides.SC_Spp_List))+
                 (base::length(Insides.AN_Spp_List))+
                 (base::length(Insides.SB_Spp_List))))
Outside_full_heat_table <-  base::data.frame(
  IslandName = Outside_Heat_Sites, Year = Outside_Heat_Year, 
  ScientificName = Outside_Heat_Species, ReserveStatus = "Outside")

Inside_full_heat_table <-  base::data.frame(
  IslandName = Inside_Heat_Sites, Year = Inside_Heat_Year, 
  ScientificName = Inside_Heat_Species, ReserveStatus = "Inside")

full_heat_table <- base::rbind(Outside_full_heat_table, Inside_full_heat_table)

heat_map <- dplyr::left_join(
  full_heat_table, heat_map_data, by=c("IslandName", "Year", "ScientificName", "ReserveStatus")) %>%
  dplyr::mutate(IslandName = dplyr::recode_factor(
    IslandName, s.SR ="Santa Rosa Island", s.SC ="Santa Cruz Island",
    s.AN ="Anacapa Island", s.SB ="Santa Barbara Island")) %>% 
  dplyr::group_by(IslandName, ScientificName, ReserveStatus) %>% 
  dplyr::mutate(frequency = length(stat[!is.na(stat)])) %>% 
  dplyr::ungroup() %>% 
  dplyr::arrange(desc(frequency), desc(ReserveStatus)) 

z <- 1
for(i in levels(heat_map$IslandName)) {
  Island_Heat_Data <- heat_map %>% 
    dplyr::filter(IslandName == i) 
  
  p <- ggplot2::ggplot() +
    geom_tile(data = Island_Heat_Data,
              aes(x = Year, y = ScientificName, fill = stat), color = "black") +
    ggplot2::scale_x_continuous(breaks = 2005:Year_to_Filter_Data_by, expand = c(0, 0)) +
    ggplot2::scale_y_discrete(expand = c(0, 0)) +
    ggplot2::scale_fill_gradient2(low = scales::muted("lightgreen"), 
                                  high = "forestgreen", na.value = scales::muted("firebrick")) + 
    ggplot2::facet_grid(rows = vars(IslandName), 
                        cols = vars(ReserveStatus), scales = "free", space = "free") +
    ggplot2::labs(title = i, x = "Year", y = "Species/Taxa", 
                  fill = paste("Strength of", "\n", "Association")) +
    ISA_theme()
  unique_plotname <- paste("ISA.", z, sep = "")
  assign(unique_plotname, p)
  z <- z + 1
}
ISA.1
ISA.2
ISA.3
ISA.4
```
